% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrels.R
\docType{data}
\name{squirrels}
\alias{squirrels}
\title{Central Park Squirrel Census, 2018}
\format{A data frame with 3023 observations on the following 17 variables.
\describe{
    \item{long}{Longitude coordinate for squirrel sighting point}
    \item{lat}{Latitude coordinate for squirrel sighting point}
    \item{unique_squirrel_id}{Identification tag for each squirrel sightings. The tag is comprised of "Hectare ID" + "Shift" + "Date" + "Hectare Squirrel Number."}
    \item{hectare}{ID tag, which is derived from the hectare grid used to divide and count the park area. One axis that runs predominantly north-to-south is numerical (1-42), and the axis that runs predominantly east-to-west is roman characters (A-I).}
    \item{shift}{Value is either "AM" or "PM," to communicate whether or not the sighting session occurred in the morning or late afternoon.}
    \item{date}{Concatenation of the sighting session day and month.}
    \item{hectare_squirrel_number}{Number within the chronological sequence of squirrel sightings for a discrete sighting session.}
    \item{age}{Value is either "Adult" or "Juvenile."}
    \item{primary_fur_color}{Value is either "Gray," "Cinnamon" or "Black."}
    \item{highlight_fur_color}{Discrete value or string values comprised of "Gray," "Cinnamon" or "Black."}
    \item{combination_of_primary_and_highlight_color}{A combination of the previous two columns; this column gives the total permutations of primary and highlight colors observed.}
    \item{color_notes}{Sighters occasionally added commentary on the squirrel fur conditions. These notes are provided here.}
    \item{location}{Value is either "Ground Plane" or "Above Ground." Sighters were instructed to indicate the location of where the squirrel was when first sighted.}
    \item{above_ground_sighter_measurement}{For squirrel sightings on the ground plane, fields were populated with a value of “FALSE.”}
    \item{specific_location}{Sighters occasionally added commentary on the squirrel location. These notes are provided here.}
    \item{running}{Squirrel was seen running.}
    \item{chasing}{Squirrel was seen chasing another squirrel.}
    \item{climbing}{Squirrel was seen climbing a tree or other environmental landmark.}
    \item{eating}{Squirrel was seen eating.}
    \item{foraging}{Squirrel was seen foraging for food.}
    \item{other_activities}{}
    \item{kuks}{Squirrel was heard kukking, a chirpy vocal communication used for a variety of reasons.}
    \item{quaas}{Squirrel was heard quaaing, an elongated vocal communication which can indicate the presence of a ground predator such as a dog.}
    \item{moans}{Squirrel was heard moaning, a high-pitched vocal communication which can indicate the presence of an air predator such as a hawk.}
    \item{tail_flags}{Squirrel was seen flagging its tail. Flagging is a whipping motion used to exaggerate squirrel's size and confuse rivals or predators. Looks as if the squirrel is scribbling with tail into the air.}
    \item{tail_twitches}{Squirrel was seen twitching its tail. Looks like a wave running through the tail, like a breakdancer doing the arm wave. Often used to communicate interest, curiosity.}
    \item{approaches}{Squirrel was seen approaching human, seeking food.}
    \item{indifferent}{Squirrel was indifferent to human presence.}
    \item{runs_from}{Squirrel was seen running from humans, seeing them as a threat.}
    \item{other_interactions}{Sighter notes on other types of interactions between squirrels and humans.}
    \item{zip_codes}{}
    \item{community_districts}{}
    \item{borough_boundaries}{}
    \item{city_council_districts}{}
    \item{police_precincts}{}
  }}
\source{
https://data.cityofnewyork.us/Environment/2018-Central-Park-Squirrel-Census-Squirrel-Data/vfnx-vebw
}
\usage{
squirrels
}
\description{
The Squirrel Census (https://www.thesquirrelcensus.com/) is a multimedia
science, design, and storytelling project focusing on the Eastern gray
(Sciurus carolinensis). They count squirrels and present their findings to
the public. This table contains squirrel data for each of the 3,023 sightings,
including location coordinates, age, primary and secondary fur color,
elevation, activities, communications, and interactions between squirrels
and with humans.
}
\keyword{datasets}
